refactor: move ensure helpers to compile packages
